refactor: restructure bots related code into modular components
