+++ components/server/server-section.tsx
@@ -55,7 +55,7 @@ export const ServerSection = ({
       onClick={() => onOpen("createChatBot", { server })}
       className="text-zinc-500 hover:text-zinc-600 dark:text-zinc-400 dark:hover:text-zinc-300 transition"
      >
-      <Plus className="h-4 w-4" />
+      {/* <Plus className="h-4 w-4" /> */}
      </button>
     </ActionTooltip>
    )}
