+++ lib/types/global.d.ts
@@ -1,0 +1,5 @@
+declare global {
+  interface Window {
+    pdfjsLib: any;
+  }
+}
